Add internal stream reading, async iteration protocol, and stream fixes
Introduce native, content-invisible and optimized ReadableStream sources
and sinks for such as fetch bodies or the File API.

While working through these changes, I made various improvements to
other parts of the codebase, which are hard to disentangle, so they
remain part of this one commit.

Most notably, the `js` crate has the following changes applied:

- Add `iteration.rs`: ECMAScript async iteration protocol support —
  `GetIterator(obj, async)` with the `CreateAsyncFromSyncIterator`
  fallback,
  `IteratorNext`, return teardown, and `CreateIterResultObject`. The
  async-from-sync wrapper caches its reaction callbacks and fast-paths
  primitive yielded values.
- Add `Promise::shared_resolved_undefined` for internal "already
  resolved"
  plumbing; rename `Promise::call_original_then` -> `Promise::then`;
  make
  `new_rejected_with_pending_error` resolve with `undefined` when no
  exception
  is pending (uncatchable exceptions).
- Add `OptionHeapExt` (`get`/`take_rooted` for `Option<Heap<T>>`) and
  `ToJSVal` for
  `Heap<T>`.
- Assorted fixes to the class/handle infrastructure.

Where required, builtins are adapted to these changes.

+++ crates/builtins/web-streams/src/readable/underlying_source.rs
@@ -13,13 +13,17 @@ pub struct UnderlyingSource<'a> {
     pub pull: Option<PullCallback<'a>>,
     pub cancel: Option<CancelCallback<'a>>,
     pub r#type: Option<ReadableStreamType>,
-    pub auto_allocate_chunk_size: Option<u64>,
+    // WebIDL `[EnforceRange] unsigned long long`. Held as `f64` so the sign and
+    // non-finiteness survive the dictionary conversion (the macro converts
+    // integers with wrapping semantics); `[EnforceRange]` is applied in
+    // `set_up_readable_byte_stream_controller_from_underlying_source`.
+    pub auto_allocate_chunk_size: Option<f64>,
 }
 
-/// WebIDL callback `UnderlyingSourceStartCallback`: (controller: ReadableStreamController<'_>) -> HandleValue<'_>
+/// WebIDL callback `UnderlyingSourceStartCallback`: (controller: ReadableStreamController) -> HandleValue<'_>
 pub type StartCallback<'s> = Object<'s>;
 
-/// WebIDL callback `UnderlyingSourcePullCallback`: (controller: ReadableStreamController<'_>) -> Promise<'_>
+/// WebIDL callback `UnderlyingSourcePullCallback`: (controller: ReadableStreamController) -> Promise<'_>
 pub type PullCallback<'s> = Object<'s>;
 
 /// WebIDL callback `UnderlyingSourceCancelCallback`: (reason: Option<HandleValue<'_>>) -> Promise<'_>
